+++ ak_library_management/wizard/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import sale_order_validation_wizard
-from . import borrow_books_wizard
+from . import borrow_books_checklist_wizard