+++ ak_library_management/changelog.md
@@ -49,3 +49,7 @@ It should not be visible if no products have been created yet.
 
 ## [Version 2.4.0]()
 ### Created borrow books wizard model and its view.
+
+## [Version 2.5.0]()
+### Created borrow books checklist wizard (views/model)
+- for showing dynamics warning